// src/users/model/notification-category.ts
export enum NotificationCategory {
  USER_RECEIVE_FUND_VIA_IN_APP_TRANSFER = 'USER_RECEIVE_FUND_VIA_IN_APP_TRANSFER',
  VENDOR_RECEIVE_FUND_VIA_IN_APP_TRANSFER = 'VENDOR_RECEIVE_FUND_VIA_IN_APP_TRANSFER',
  USER_RECEIVE_FUND_VIA_PAYSTACK_WEBHOOK = 'USER_RECEIVE_FUND_VIA_PAYSTACK_WEBHOOK',
  USER_CANCELLED_ORDER = 'USER_CANCELLED_ORDER',
  VENDOR_RECEIVE_FUND_VIA_PAYSTACK_WEBHOOK = 'VENDOR_RECEIVE_FUND_VIA_PAYSTACK_WEBHOOK',
  VENDOR_RECEIVE_ORDER = 'VENDOR_RECEIVE_ORDER',
  ORDER_REFUND = 'ORDER_REFUND',
  VENDOR_ORDER_REWARD = 'VENDOR_ORDER_REWARD',
  ORDER_PICKED_UP = 'ORDER_PICKED_UP',
  ORDER_CANCELLED = 'ORDER_CANCELLED',
  ORDER_DELIVERED = 'ORDER_DELIVERED',
  VENDOR_CREDENTIAL_APPROVED = 'VENDOR_CREDENTIAL_APPROVED',
  VENDOR_CREDENTIAL_REJECTED = 'VENDOR_CREDENTIAL_REJECTED',
  VENDOR_REJECT_ORDER = 'VENDOR_REJECT_ORDER',
  VENDOR_ACCEPT_ORDER = 'VENDOR_ACCEPT_ORDER',
  DOCTOR_NEW_APPOINTMENT = 'DOCTOR_NEW_APPOINTMENT',
  APPOINTMENT_BOOKED = 'APPOINTMENT_BOOKED',
  APPOINTMENT_REMINDER = 'APPOINTMENT_REMINDER',
  DOCTOR_APPOINTMENT_REMINDER = 'DOCTOR_APPOINTMENT_REMINDER',
  APPOINTMENT_REJECTED = 'APPOINTMENT_REJECTED',
  APPOINTMENT_COMPLETED = 'APPOINTMENT_COMPLETED',
  DOCTOR_VERIFICATION_STATUS_UPDATED = 'DOCTOR_VERIFICATION_STATUS_UPDATED',
  RIDER_TRAINING_COMPLETED = 'RIDER_TRAINING_COMPLETED',
  RIDER_DOCUMENT_VERIFIED = 'RIDER_DOCUMENT_VERIFIED',
  RIDER_DOCUMENT_REJECTED = 'RIDER_DOCUMENT_REJECTED',
  RIDER_BACKGROUND_CHECK_APPROVED = 'RIDER_BACKGROUND_CHECK_APPROVED',
  RIDER_BACKGROUND_CHECK_REJECTED = 'RIDER_BACKGROUND_CHECK_REJECTED',
  RIDER_BIKE_REGISTERED = 'RIDER_BIKE_REGISTERED',
  RIDER_ONLINE_STATUS_CHANGED = 'RIDER_ONLINE_STATUS_CHANGED',
  RIDER_STATUS_CHANGED = 'RIDER_STATUS_CHANGED',
  IMMEDIATE_CONSULTATION_REQUEST = 'IMMEDIATE_CONSULTATION_REQUEST',
  CONSULTATION_STARTED = 'CONSULTATION_STARTED',
  CONSULTATION_REJECTED = 'CONSULTATION_REJECTED',
  CONSULTATION_COMPLETED = 'CONSULTATION_COMPLETED',
  CONSULTATION_CANCELLED = 'CONSULTATION_CANCELLED',
  CONSULTATION_EXPIRED = 'CONSULTATION_EXPIRED',
  NEW_ORDER_AVAILABLE = 'NEW_ORDER_AVAILABLE',
  BIRTHDAY_REWARD = 'BIRTHDAY_REWARD',
}

export const NotificationRoutes = {
  USER_RECEIVE_FUND_VIA_IN_APP_TRANSFER: '/(wallet)/TransactionHistory',
  USER_RECEIVE_FUND_VIA_PAYSTACK_WEBHOOK: '/(wallet)/TransactionHistory',
  VENDOR_RECEIVE_FUND_VIA_IN_APP_TRANSFER:
    '/vendor/(wallet)/TransactionHistory',
  VENDOR_RECEIVE_FUND_VIA_PAYSTACK_WEBHOOK:
    '/vendor/(wallet)/TransactionHistory',
  VENDOR_RECEIVE_ORDER: '/vendor/(tabs)',
  ORDER_REFUND: '/(tabs)/wallet',
  VENDOR_ORDER_REWARD: '/(vendor)/wallet',
  ORDER_PICKED_UP: '/(tabs)/order',
  ORDER_CANCELLED: '/(tabs)/order',
  ORDER_DELIVERED: '/(tabs)/order',
  ORDER_STATUS: '/(tabs)/order',
  VENDOR_CREDENTIAL_APPROVED: '/(vendor)/credentials', //TODO: Update this route to point to the vendor credentials page
  VENDOR_CREDENTIAL_REJECTED: '/(vendor)/credentials',
  IMMEDIATE_CONSULTATION_REQUEST: '/Doctors/(tabs)',
  DOCTOR_NEW_APPOINTMENT: '/Doctors/(tabs)?tab=scheduled',
  APPOINTMENT_REMINDER: '/(menu)/(medtechTabs)/Appointment?tab=upcoming',
  DOCTOR_APPOINTMENT_REMINDER: '/Doctors/(tabs)?tab=scheduled',
  APPOINTMENT_REJECTED: '/(menu)/(medtechTabs)/Appointment?tab=past',
  CONSULTATION_STARTED: '/(medicalAi)/Meeting',
  CONSULTATION_CANCELLED: '/Doctors/(tabs)?tab=cancelled',
  CONSULTATION_EXPIRED: '/(menu)/(medtechTabs)/Appointment?tab=past',
  CONSULTATION_COMPLETED: '/(doc-consult)/ConsultationSummary',
  RIDER_VERIFICATION: 'RIDER_VERIFICATION',
  BIRTHDAY_REWARD: '/(tabs)',
};